requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1

# Database & Cache
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
alembic==1.13.0

# Data Processing & ML
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
scipy==1.11.4
joblib==1.3.2
xgboost==2.0.1
lightgbm==4.1.0

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.1.0
mediapipe==0.10.8
face-recognition==1.3.0
dlib==19.24.2

# AI & NLP
openai==1.3.7
transformers==4.36.2
torch==2.1.1
torchvision==0.16.1

# Security & Authentication
PyJWT==2.8.0
bcrypt==4.1.1
passlib[bcrypt]==1.7.4
cryptography==41.0.8

# Background Tasks
celery==5.3.4
flower==2.0.1

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk==1.38.0

# Validation & Serialization
marshmallow==3.20.1
email-validator==2.1.0

# HTTP & Networking
httpx==0.25.2
requests==2.31.0
websockets==12.0

# Configuration
python-decouple==3.8
pyyaml==6.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7

# Frontend (if using Streamlit)
streamlit==1.28.1
plotly==5.17.0
dash==2.14.2

# File Processing
openpyxl==3.1.2
python-docx==1.1.0
pypdf2==3.0.1